refactor: update imports for CompareTotals and related types
