+++ client/src/lib/types/compare.ts
@@ -1,3 +1,6 @@
+import type { ScenarioTotals } from '$lib/process-results/processCases';
+import type { EmulatorResults, Scenario } from './userState';
+
 export interface CompareParameter {
 	parameterName: string;
 	label: string;
@@ -16,4 +19,16 @@ export interface InterventionCompareParameter extends CompareParameter {
 export interface CompareParameters {
 	baselineParameters: CompareParameter[];
 	interventionParameters: InterventionCompareParameter[];
+}
+
+export type CompareSeriesName = 'Present' | 'Long term (baseline + control strategy)' | 'Long term (baseline only)';
+export interface CompareResults {
+	present: EmulatorResults;
+	fullLongTerm: EmulatorResults;
+	baselineLongTerm: EmulatorResults;
+}
+export interface CompareTotals {
+	presentTotals: Partial<Record<Scenario, ScenarioTotals>>;
+	baselineLongTermTotals: Partial<Record<Scenario, ScenarioTotals>>;
+	fullLongTermTotals: Partial<Record<Scenario, ScenarioTotals>>;
 }
